Shut up some compiler warnings (some warnings should be enabled again in autoconfig_msvc.h IMHO because hiding our possible mistakes won't help us).
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@39003 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/common/classes/TempFile.cpp
@@ -147,15 +147,13 @@ void TempFile::init(const Firebird::PathName& directory, const Firebird::PathNam
 	{
 		Firebird::PathName name = filename + prefix;
 		__int64 temp = randomness;
-		for (int i = 0; i < suffix.length(); i++)
+		for (size_t i = 0; i < suffix.length(); i++)
 		{
 			suffix[i] = NAME_LETTERS[temp % (strlen(NAME_LETTERS))];
 			temp /= strlen(NAME_LETTERS);
 		}
 		name += suffix;
-		DWORD attributes = FILE_ATTRIBUTE_NORMAL |
-						   FILE_ATTRIBUTE_TEMPORARY |
-						   FILE_FLAG_RANDOM_ACCESS;
+		DWORD attributes = FILE_ATTRIBUTE_NORMAL | FILE_ATTRIBUTE_TEMPORARY | FILE_FLAG_RANDOM_ACCESS;
 		if (doUnlink)
 		{
 			attributes |= FILE_FLAG_DELETE_ON_CLOSE;
